refactor(intersection): cleanup utility function (#5972)
